# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: API Minha Folga
  version: 1.1.0
  summary: Cadastro de interesse, preferências, atendimento, ferramentas da Bia (Hal-AI) e administração.
  description: |
    Contrato versionado entre `frontend/` (Vue) e `backend/` (Node/Express). Todas as rotas vivem sob `/api`,
    na mesma origem do site. Regras gerais:

    - Erros sempre em JSON: `{ "error": { "code", "message", "fields"? , ...extra }, "requestId" }`.
      `message` é texto em português que pode ser exibido ao usuário; `code` é estável.
    - Chamadas do navegador com método inseguro exigem `Content-Type: application/json` e, se houver
      cabeçalho `Origin`, que ela esteja na lista permitida (defesa CSRF de base).
    - Corpos com campos inesperados são recusados (`validation_error`). Limite de corpo: 16 KB em todas as rotas.
    - Web Push usa autorização independente de navegador, endpoint cifrado e token técnico do dono; ver notifications.openapi.yaml.
    - Sessões usam cookies `HttpOnly; Secure; SameSite=Strict` definidos pelo servidor. Nenhum token vai para
      `localStorage`, query string ou pixels. Dados pessoais nunca aparecem em URL.
    - Respostas de cadastro e de pedido de saída são genéricas: não revelam se um número ou CPF já está cadastrado.
    - Em `PRE_LAUNCH`, toda rota `/api/credit/*` responde 403 `credit_phase_locked`.

    Decisão D1 (docs/DECISOES.md, 26/09/2026): este sistema **não mexe com WhatsApp**. Número oficial, Bia, webchat,
...
servers:
...
paths:
...
  /api/support:
    post:
      tags: [ atendimento ]
      summary: Registra solicitação de atendimento, gera protocolo aleatório e encaminha à fila interna.
      description: |
        A solicitação é persistida antes de qualquer notificação; falha de notificação não apaga o registro.
        O `trackingToken` é devolvido uma única vez e permite acompanhar o status; o protocolo sozinho não autoriza leitura.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SupportRequest' }
      responses:
        '201': { description: Solicitação registrada., content: { application/json: { schema: { $ref: '#/components/schemas/SupportCreated' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { description: '`support_unavailable`.', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/support/{protocol}:
    get:
      tags: [ atendimento ]
      summary: Consulta o status de uma solicitação. Exige o token de acompanhamento do titular.
      parameters:
        - { in: path, name: protocol, required: true, schema: { type: string, pattern: '^MF-[0-9A-Z]{4}-[0-9A-Z]{4}-[0-9A-Z]{2}$' } }
        - { in: header, name: X-Support-Token, required: true, schema: { type: string } }
      responses:
        '200': { description: Status da solicitação., content: { application/json: { schema: { $ref: '#/components/schemas/SupportStatus' } } } }
        '404': { description: 'Protocolo inexistente ou token inválido (mesma resposta: `not_found`).', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '429': { $ref: '#/components/responses/RateLimited' }

  # --------------------------------------------------------------------------------------------- métricas
...
  /api/admin/support:
    get:
      tags: [ admin ]
      summary: Fila de atendimento (support:read).
      description: Cada consulta é auditada (`admin.support_listed`, com filtros, página e quantidade; sem dados do titular).
      parameters:
        - { in: query, name: status, schema: { $ref: '#/components/schemas/SupportState' } }
        - { in: query, name: assignee, schema: { type: string, description: 'uuid, "me" ou "none"' } }
        - { in: query, name: from, schema: { type: string, format: date } }
        - { in: query, name: to, schema: { type: string, format: date } }
        - { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
        - { in: query, name: pageSize, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
      responses:
        '200':
          description: Página da fila.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/AdminSupportItem' } }
                  total: { type: integer }
                  page: { type: integer }
                  pageSize: { type: integer }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/admin/support/{id}:
    get:
      tags: [ admin ]
      summary: Detalhe com mensagem e contato decifrados (support:read); acesso auditado.
      parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
      responses:
        '200': { description: Detalhe., content: { application/json: { schema: { $ref: '#/components/schemas/AdminSupportDetail' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [ admin ]
      summary: Altera status e/ou responsável (support:update). Registra responsável e horário.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
        - { $ref: '#/components/parameters/Csrf' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                status: { $ref: '#/components/schemas/SupportState' }
                assigneeId: { type: [ string, 'null' ], format: uuid }
      responses:
        '200': { description: Atualizado., content: { application/json: { schema: { $ref: '#/components/schemas/AdminSupportDetail' } } } }
        '400': { description: '`invalid_transition` ou validação.', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    adminSession: { type: apiKey, in: cookie, name: mf_admin }
    contactSession: { type: apiKey, in: cookie, name: mf_contact }
    agentApiKey: { type: apiKey, in: header, name: X-API-Key }
  parameters:
    Csrf:
      in: header
      name: X-CSRF-Token
      required: true
      description: Token da sessão administrativa (de /api/admin/auth/me ou do login). Obrigatório em métodos inseguros do admin.
      schema: { type: string }
  responses:
    ValidationError:
      description: '`validation_error` com `fields` { campo: mensagem }.'
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    RateLimited:
      description: '`rate_limited`.'
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    Forbidden:
      description: '`forbidden` (sem permissão; tentativa auditada).'
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: '`not_found`.'
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    Error:
      type: object
      required: [ error ]
      properties:
        error:
          type: object
          required: [ code, message ]
          properties:
            code: { type: string }
            message: { type: string }
            fields: { type: object, additionalProperties: { type: string } }
            attemptsRemaining: { type: integer }
            retryAfterSeconds: { type: integer }
        requestId: { type: string }
...
    SupportState: { type: string, enum: [ received, in_progress, answered, closed ] }
    AdminRole: { type: string, enum: [ admin, support, privacy, marketing ] }
...
    SupportRequest:
      type: object
      additionalProperties: false
      required: [ name, contactMethod, contact, subject, message ]
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
        contactMethod: { type: string, enum: [ email, whatsapp ] }
        contact: { type: string, maxLength: 120, description: 'E-mail ("Confira o e-mail.") ou celular com DDD ("Confira o DDD e o número.").' }
        subject: { type: string, description: 'Uma de contracts/validation.json → support.subjects.' }
        message: { type: string, minLength: 10, maxLength: 2000 }
    SupportCreated:
      type: object
      required: [ protocol, trackingToken, status, message ]
      properties:
        protocol: { type: string, example: MF-7K3P-9QXA-2M }
        trackingToken: { type: string, description: Exibido uma única vez; compõe o link seguro de acompanhamento (fragmento da URL). }
        status: { const: received }
        message: { type: string }
    SupportStatus:
      type: object
      properties:
        protocol: { type: string }
        status: { $ref: '#/components/schemas/SupportState' }
        subject: { type: string }
        subjectLabel: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        timeline:
          type: array
          items:
            type: object
            properties:
              type: { type: string, enum: [ created, status_changed, reply ] }
              status: { $ref: '#/components/schemas/SupportState' }
              text: { type: [ string, 'null' ], description: Somente respostas marcadas como visíveis ao titular. }
              at: { type: string, format: date-time }
...
    AdminUserRef:
      type: object
      properties:
        id: { type: string, format: uuid }
        displayName: { type: string }
        role: { $ref: '#/components/schemas/AdminRole' }
...
    AdminSupportItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        protocol: { type: string }
        subject: { type: string }
        status: { $ref: '#/components/schemas/SupportState' }
        source: { type: string, enum: [ web_form, agent ] }
        contactMethod: { type: string }
        contactHint: { type: string }
        assignee: { oneOf: [ { $ref: '#/components/schemas/AdminUserRef' }, { type: 'null' } ] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AdminSupportDetail:
      allOf:
        - { $ref: '#/components/schemas/AdminSupportItem' }
        - type: object
          properties:
            requesterName: { type: string }
            contact: { type: string }
            message: { type: string }
            firstResponseAt: { type: [ string, 'null' ], format: date-time }
            closedAt: { type: [ string, 'null' ], format: date-time }
            events:
              type: array
              items:
                type: object
                properties:
                  type: { type: string }
                  fromStatus: { type: [ string, 'null' ] }
                  toStatus: { type: [ string, 'null' ] }
                  note: { type: [ string, 'null' ] }
                  visibleToRequester: { type: boolean }
                  actor: { oneOf: [ { $ref: '#/components/schemas/AdminUserRef' }, { type: 'null' } ] }
                  actorType: { type: string }
                  at: { type: string, format: date-time }
                  assignee:
                    description: Responsável naquele momento (eventos de atribuição).
                    oneOf:
                      - $ref: "#/components/schemas/AdminUserRef"
                      - type: "null"
...